Add open hours REST endpoint
