+++ packages/features/src/index.ts
@@ -4,6 +4,7 @@ export * from "./audit-logs/list-audit-logs";
 export * from "./audit-logs/logger";
 export * from "./config";
 export * from "./db/locking";
+export * from "./open-hours";
 export * from "./periods/access";
 export * from "./periods/generate";
 export * from "./sessions";
