+++ packages/features/src/open-hours/index.ts
@@ -1,0 +1,1 @@
+export * from "./open-hours";
